app/requirements/development.txt
selenium==4.7.2
PyAutoGUI==0.9.54
python-dotenv==1.0.0